Use the LLDB that we built
Up to now, we were running our tests using the LLDB from inside the NDK.
From this change forwards, we are using the LLDB we just built for the
Linux-x86_64 host.

We are already using the lldb-server for Android-arm64 that we built
ourselves, so this means we are testing things fully, end-to-end.